Make SDL_config.h redirect to an architecture-specific version
This avoids conflicts between different-architecture versions of
libsdl2-dev, without requiring a patch to sdl2-config to make it use a
different ${includedir} per architecture. It relies on the fact that the
architecture-specific location /usr/include/${multiarch} is already on
Debian compilers' default search paths, at the cost of adding a
Debian-specific forwarding header.

This is similar to the technique used for multilib in Fedora.

Signed-off-by: Simon McVittie <smcv@debian.org>
Closes: #909740
Closes: #951087
Closes: #951943
Closes: #951959
Closes: #951974
Closes: #951976
Closes: #952049