Install all headers in an architecture-specific location
This approach was suggested by Adrian Bunk on #909740, as one of several
possible ways to avoid SDL_config.h colliding between architectures.

Because this approach uses a different @includedir@ for each architecture,
we can no longer have @includedir@ in sdl2-config.in: that would result
in non-co-installability. However, simply removing it (similar to
no-libdir.patch) is not an option, because the SDL 2 headers are not
on the default include path (and they cannot be added to it without
breaking parallel-installability of SDL 1.2 or a hypothetical future
SDL 3). Instead, patch sdl-config.in to ask pkg-config for its CFLAGS
and LIBS. This can supersede the previous no-libdir.patch.

This relies on the assumption that cross-compilation will always set
PKG_CONFIG, even if that environment variable is not otherwise used by
the build system that is invoking sdl2-config.

The sdl2-config.cmake.in part of no-libdir.patch can simply be dropped,
because that file is installed to an architecture-specific location
anyway, namely ${libdir}/cmake.

Signed-off-by: Simon McVittie <smcv@debian.org>