d/rules: Provide compatibility symlinks to begin_code.h, close_code.h
If a project that (incorrectly?) #includes <SDL2/SDL_foo.h>, such as
jag_0.3.5-5, is compiled with gcc, our symlink to SDL_platform.h is
sufficient to make it work. However, when compiled with clang, that
workaround is not sufficient. It appears that clang's behaviour when
foo.h has #include "bar.h", and foo.h is a symlink, is not the same
as gcc's: gcc dereferences the symlink and looks for bar.h in the same
directory as the regular file foo.h, but clang looks for bar.h in the
directory where it found the foo.h symlink.

Work around this by providing symlinks to everything that is required
by SDL_config.h, recursively.