Install upstream's tests
Most of them are manual tests rather than automated tests, so we don't
run them at build-time.

Package them in the -dev package for now, to avoid a trip through
the NEW queue to add a libsdl2-tests package. The installed location
is loosely based on what GNOME installed-tests do, but must be multiarch
until/unless we separate out a libsdl2-tests package, so put it below
the $(libdir) rather than $(libexecdir).

Signed-off-by: Simon McVittie <smcv@debian.org>